+++ surf-roleplay-api/src/main/kotlin/dev/slne/surf/roleplay/api/transaction/HasTransactions.kt
@@ -12,7 +12,7 @@ interface HasTransactions {
      * @param balanceType The type of balance to retrieve.
      * @return The balance amount as a [Double].
      */
-    suspend fun getBalance(balanceType: BalanceType): Double
+    suspend fun getBalance(balanceType: BalanceType): Int
 
     /**
      * Adds the specified [amount] to the player's balance of the given [BalanceType].
@@ -21,7 +21,7 @@ interface HasTransactions {
      * @param amount The amount to add to the balance.
      * @return `true` if the operation was successful, `false` otherwise.
      */
-    suspend fun addBalance(balanceType: BalanceType, amount: Double): Boolean
+    suspend fun addBalance(balanceType: BalanceType, amount: Int): Boolean
 
     /**
      * Removes the specified [amount] from the player's balance of the given [BalanceType].
@@ -30,7 +30,7 @@ interface HasTransactions {
      * @param amount The amount to remove from the balance.
      * @return `true` if the operation was successful, `false` otherwise.
      */
-    suspend fun removeBalance(balanceType: BalanceType, amount: Double): Boolean
+    suspend fun removeBalance(balanceType: BalanceType, amount: Int): Boolean
 
     /**
      * Retrieves the balance history for the specified [BalanceType].
@@ -53,15 +53,15 @@ interface HasTransactions {
      */
     suspend fun hasBalance(
         balanceType: BalanceType,
-        amount: Double
+        amount: Int
     ): Boolean = getBalance(balanceType) >= amount
 
     /**
      * Retrieves the player's cash balance.
      *
-     * @return The cash balance as a [Double].
+     * @return The cash balance as a [Int].
      */
-    suspend fun getCashBalance(): Double = getBalance(BalanceType.CASH)
+    suspend fun getCashBalance(): Int = getBalance(BalanceType.CASH)
 
     /**
      * Adds the specified [amount] to the player's cash balance.
@@ -69,7 +69,7 @@ interface HasTransactions {
      * @param amount The amount to add to the cash balance.
      * @return `true` if the operation was successful, `false` otherwise.
      */
-    suspend fun addCashBalance(amount: Double) = addBalance(BalanceType.CASH, amount)
+    suspend fun addCashBalance(amount: Int) = addBalance(BalanceType.CASH, amount)
 
     /**
      * Removes the specified [amount] from the player's cash balance.
@@ -77,7 +77,7 @@ interface HasTransactions {
      * @param amount The amount to remove from the cash balance.
      * @return `true` if the operation was successful, `false` otherwise.
      */
-    suspend fun removeCashBalance(amount: Double) = removeBalance(BalanceType.CASH, amount)
+    suspend fun removeCashBalance(amount: Int) = removeBalance(BalanceType.CASH, amount)
 
     /**
      * Checks if the player has a sufficient cash balance for the given [amount].
@@ -85,7 +85,7 @@ interface HasTransactions {
      * @param amount The amount to check against the player's cash balance.
      * @return `true` if the player has enough cash balance, `false` otherwise.
      */
-    suspend fun hasCashBalance(amount: Double): Boolean = hasBalance(BalanceType.CASH, amount)
+    suspend fun hasCashBalance(amount: Int): Boolean = hasBalance(BalanceType.CASH, amount)
 
     /**
      * Retrieves the cash balance history.
@@ -99,9 +99,9 @@ interface HasTransactions {
     /**
      * Retrieves the player's bank balance.
      *
-     * @return The bank balance as a [Double].
+     * @return The bank balance as a [Int].
      */
-    suspend fun getBankBalance(): Double = getBalance(BalanceType.BANK)
+    suspend fun getBankBalance(): Int = getBalance(BalanceType.BANK)
 
     /**
      * Transfers the specified [amount] from the player's bank balance to another [RpPlayer]'s bank balance.
@@ -110,7 +110,7 @@ interface HasTransactions {
      * @param amount The amount to transfer.
      * @return `true` if the transfer was successful, `false` otherwise.
      */
-    suspend fun transferBankBalance(receiver: RpPlayer, amount: Double): Boolean
+    suspend fun transferBankBalance(receiver: RpPlayer, amount: Int): Boolean
 
     /**
      * Adds the specified [amount] to the player's bank balance.
@@ -118,7 +118,7 @@ interface HasTransactions {
      * @param amount The amount to add to the bank balance.
      * @return `true` if the operation was successful, `false` otherwise.
      */
-    suspend fun addBankBalance(amount: Double) = addBalance(BalanceType.BANK, amount)
+    suspend fun addBankBalance(amount: Int) = addBalance(BalanceType.BANK, amount)
 
     /**
      * Removes the specified [amount] from the player's bank balance.
@@ -126,7 +126,7 @@ interface HasTransactions {
      * @param amount The amount to remove from the bank balance.
      * @return `true` if the operation was successful, `false` otherwise.
      */
-    suspend fun removeBankBalance(amount: Double) = removeBalance(BalanceType.BANK, amount)
+    suspend fun removeBankBalance(amount: Int) = removeBalance(BalanceType.BANK, amount)
 
     /**
      * Checks if the player has a sufficient bank balance for the given [amount].
@@ -134,7 +134,7 @@ interface HasTransactions {
      * @param amount The amount to check against the player's bank balance.
      * @return `true` if the player has enough bank balance, `false` otherwise.
      */
-    suspend fun hasBankBalance(amount: Double): Boolean = hasBalance(BalanceType.BANK, amount)
+    suspend fun hasBankBalance(amount: Int): Boolean = hasBalance(BalanceType.BANK, amount)
 
     /**
      * Retrieves the bank balance history.
@@ -148,9 +148,9 @@ interface HasTransactions {
     /**
      * Retrieves the player's crypto balance.
      *
-     * @return The crypto balance as a [Double].
+     * @return The crypto balance as a [Int].
      */
-    suspend fun getCryptoBalance(): Double = getBalance(BalanceType.CRYPTO)
+    suspend fun getCryptoBalance(): Int = getBalance(BalanceType.CRYPTO)
 
     /**
      * Adds the specified [amount] to the player's crypto balance.
@@ -158,7 +158,7 @@ interface HasTransactions {
      * @param amount The amount to add to the crypto balance.
      * @return `true` if the operation was successful, `false` otherwise.
      */
-    suspend fun addCryptoBalance(amount: Double) = addBalance(BalanceType.CRYPTO, amount)
+    suspend fun addCryptoBalance(amount: Int) = addBalance(BalanceType.CRYPTO, amount)
 
     /**
      * Removes the specified [amount] from the player's crypto balance.
@@ -166,7 +166,7 @@ interface HasTransactions {
      * @param amount The amount to remove from the crypto balance.
      * @return `true` if the operation was successful, `false` otherwise.
      */
-    suspend fun removeCryptoBalance(amount: Double) = removeBalance(BalanceType.CRYPTO, amount)
+    suspend fun removeCryptoBalance(amount: Int) = removeBalance(BalanceType.CRYPTO, amount)
 
     /**
      * Checks if the player has a sufficient crypto balance for the given [amount].
@@ -174,7 +174,7 @@ interface HasTransactions {
      * @param amount The amount to check against the player's crypto balance.
      * @return `true` if the player has enough crypto balance, `false` otherwise.
      */
-    suspend fun hasCryptoBalance(amount: Double): Boolean = hasBalance(BalanceType.CRYPTO, amount)
+    suspend fun hasCryptoBalance(amount: Int): Boolean = hasBalance(BalanceType.CRYPTO, amount)
 
     /**
      * Retrieves the crypto balance history.
